* commit '089b44acfae33f5f9aafe701b0e3c059cb316339': (27 commits)
  系统更新接口重复提示bug
  更新发布版本号
  系统更新接口时间格式修改
  换图片教程；
  消息页小数
  智能下单；常购清单接口数量小数
  修改手机toast；出库页可输入小数
  退货小数；修改密码
  修改密码
  调拨小数
  修改密码后退出并提示
  收货有批次
  收货，无批次小数
  轮播栏修改
  去掉用户手册错误的测试代码
  为指纹识别修改重置密码页
  购物车小数
  常购清单和智能下单完成，接口商品信息通过product字段传
  下单使用小数修改成功
  取消指纹登录也要验证指纹；修改密码后也修改指纹登录的密码
  ...

+++ app/src/main/java/com/runwise/supply/entity/TransferBatchLot.java
@@ -11,19 +11,19 @@ import android.os.Parcelable;
 
 public class TransferBatchLot implements Parcelable{
     private String lotID;
-    private int quantQty;
+    private double quantQty;
     private String lotDate;
     private int lotIDID;
     private float usedQty;
 
     //本地数据
-    private int actualQty;
+    private double actualQty;
 
     public String getLotID() {
         return lotID;
     }
 
-    public int getQuantQty() {
+    public double getQuantQty() {
         return quantQty;
     }
 
@@ -35,11 +35,11 @@ public class TransferBatchLot implements Parcelable{
         this.quantQty = quantQty;
     }
 
-    public int getActualQty() {
+    public double getActualQty() {
         return actualQty;
     }
 
-    public void setActualQty(int actualQty) {
+    public void setActualQty(double actualQty) {
         this.actualQty = actualQty;
     }
 
@@ -78,8 +78,8 @@ public class TransferBatchLot implements Parcelable{
     @Override
     public void writeToParcel(Parcel dest, int flags) {
         dest.writeString(this.lotID);
-        dest.writeInt(this.quantQty);
-        dest.writeInt(this.actualQty);
+        dest.writeDouble(this.quantQty);
+        dest.writeDouble(this.actualQty);
         dest.writeFloat(this.usedQty);
         dest.writeString(this.lotDate);
         dest.writeInt(this.lotIDID);
@@ -87,8 +87,8 @@ public class TransferBatchLot implements Parcelable{
 
     protected TransferBatchLot(Parcel in) {
         this.lotID = in.readString();
-        this.quantQty = in.readInt();
-        this.actualQty = in.readInt();
+        this.quantQty = in.readDouble();
+        this.actualQty = in.readDouble();
         this.usedQty = in.readFloat();
         this.lotDate = in.readString();
         this.lotIDID = in.readInt();
