+++ app/src/main/java/com/runwise/supply/firstpage/entity/ReturnRequest.java
@@ -25,15 +25,15 @@ public class ReturnRequest {
          * reason : xxx
          */
 
-        private int qty;
+        private double qty;
         private int product_id;
         private String reason;
 
-        public int getQty() {
+        public double getQty() {
             return qty;
         }
 
-        public void setQty(int qty) {
+        public void setQty(double qty) {
             this.qty = qty;
         }
